apps/api/pyproject.toml
[project]
name = "licitscope-api"
version = "0.1.0"
description = "LicitScope – AI procurement intelligence for Brazil (backend)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "LicitScope contributors" }]

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlmodel>=0.0.22",
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    "httpx>=0.27",
    "python-dateutil>=2.9",
    "rapidfuzz>=3.9",
    "unidecode>=1.3",
    "tenacity>=9.0",
]

[project.optional-dependencies]
postgres = ["psycopg[binary]>=3.2"]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "httpx>=0.27",
    "respx>=0.21",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "RUF"]
# UP042: don't insist on StrEnum — we want JSON values unchanged on the wire
# RUF012: SQLAlchemy columns as class attributes is idiomatic here
ignore = ["E501", "B008", "N818", "UP042", "RUF012"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra -q"
filterwarnings = [
    "ignore::DeprecationWarning",
]
